Rename QConcurrentCall to QThreadFunction
The strict definition of concurrency means it needs to exist
some concurrency for resources between tasks.

This is a misleading term for the QConcurrentCall,
and most probably for the whole QtConcurrent module.

So, be more precise and rename the class into what is it doing exactly.

Rename setConcurrentCallData() into setThreadFunctionData().

Fixes: QTBUG-142924
Change-Id: I41094394144dcc9e642639750cebe768769ac62b
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 0e2287748a4cba105b75942da968bc001d1668cc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>