Rename QConcurrentCall to QThreadFunction
The strict definition of concurrency means it needs to exist
some concurrency for resources between tasks.

This is a misleading term for the QConcurrentCall,
and most probably for the whole QtConcurrent module.

So, be more precise and rename the class into what is it doing exactly.

Rename setConcurrentCallData() into setThreadFunctionData().

Pick-to: 6.11
Fixes: QTBUG-142924
Change-Id: I41094394144dcc9e642639750cebe768769ac62b
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>